requirements.txt
pillow
piexif
tqdm
iptcinfo3
